requirements.txt
# Core dependencies
click>=8.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.7.0
pillow>=10.2.0

# Optional dependencies for advanced features
ascii-magic>=2.3.0
aiofiles>=23.2.0
fastapi>=0.109.0
uvicorn>=0.27.0  # For FastAPI webhook server

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0